max_fd is updated in state applier
